Create Grammar_v2
This is as far as I have gotten with the grammar so far. It looks like we will have to do some semantic checks to ensure table names, database names, and field names exist in their proper place. Not sure how this will work with flex and bison as I have heard the c source code is pretty rough on the eyes